+++ src/components/Editor/Sidebar.vue
