+++ src/components/Editor/Toolbar.vue
@@ -143,7 +143,7 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import { EventBus } from '../main';
+import { EventBus } from './../../main';
 
 export default Vue.extend({
   data() {
