+++ src/components/HomePage/DesignChoice.vue
